Task1 fix

+++ class2_homework/Martin_Stoyanov_1.rb
@@ -4,7 +4,7 @@ income = 0.00
 expense = 0.00
 
 CSV.foreach("bank.csv") do |row|
-    if(row[0].to_i>=ARGV[0].to_i) && (row[0].to_i<=ARGV[1].to_i)
+    if(row[0].to_f>=ARGV[0].to_f) && (row[0].to_f<=ARGV[1].to_f)
         income=income+row[1].to_f
         expense=expense+row[2].to_f
     end
